+++ CHANGELOG.md
@@ -2,9 +2,8 @@
 
 # 1.2.1 
 
-### Added
- - Exported data constructors instance for `Plutarch.Lift (LiftError)`
-
+- Exported data constructors instance for `Plutarch.Lift (LiftError)`
+ 
 # 1.2 (WIP changelog)
 
 - Changed fields of `PTxInfo` whose only representation is data to not be wrapped by `PAsData`.